Added login box/button - needs to be linked to Login.cshtml form
Better chart loading

VP now remembers where you were scrolled to when you refresh and loads you up in the same position. This annoyingly dosen't work for what it was supposed to work for - ie when you vote your still pushed to the top of the page - looks a bit crappy.